pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "micro_scope"
version = "1.0.0"
description = "Market microstructure analysis tool for high-performance trading"
authors = [
  { name = "Finley Sherwell", email = "finleysherwell@icloud.com" }
]
requires-python = ">=3.8"

dependencies = [
  "numpy>=2.2.4",
  "numba>=0.61.2",
  "pandas>=2.2.3",
  "matplotlib>=3.10.1",
  "seaborn>=0.13.2",
  "requests>=2.32.3",
  "PyYAML>=6.0.2",
  "streamlit>=1.44.1",
  "dotenv>=0.9.9"
]

[project.scripts]
microscope = "micro_scope.cli:main"
